Version 2.0 of Study Guide Creator - COMPLETE VERSION
This is the first completed version of the Study Guide Creator! Run this code in a IDE or complier and create your own study guide with the output to study for a test. So far, it only supports 20 questions. (I'll add more later). It will ask you for the amount of questions you would like to have, what the questions should be, and what the answers should be. Then, it will let you answer them. If you get the question right, it will tell you and give you a point. If you get it wrong, it will tell you and give you the correct answer. Then, at the end, it will give you your score. (The score is how many questions you got correct.)