Allows for a unlimited amount of questions and some more.
I added a lot of new features (and the code is a LOT shorter). The main feature is that it allows for a unlimited amount of questions. Also, if the amount of questions is equal to 0, it prints a error message and terminates the program. And, if the score is equal to one, the text that it prints uses the correct grammar.